art: Refactor the code, Move functions into utils folder, and make it reusable.
