+++ utils/swearFilter.ts
@@ -3,7 +3,7 @@ import Filter from 'bad-words';
 // Make a global filter object instead of instantiating a new one everytime
 const filter = new Filter();
 
-export default function filterBadWords(text: string | undefined): string {
+export default function swearFilter(text: string | undefined): string {
   if (text === '' || !text) return;
 
   text = text
